+++ admin/form/forms.php
@@ -1116,8 +1116,8 @@ class Enlimbo_Forms_Aione
             if ( $element['#type'] == 'file' ) {
                 return $_FILES[$name]['tmp_name'];
             }
-            return isset( $_REQUEST[$name] ) ? sanitize_text_field( $_REQUEST[$name] ) : in_array( $element['#type'],
-                            array('textfield', 'textarea') ) ? '' : 0;
+            return isset( $_REQUEST[$name] ) ? sanitize_text_field( $_REQUEST[$name] ) : (in_array( $element['#type'],
+                            array('textfield', 'textarea') ) ? '' : 0);
         }
 
         if ( !function_exists('getSubmittedDataTrim')) {